replace adding statsd directly and use an instrumenter instead
Intead of passing an stats client and calling stats directly from the transporter, we can instrument all of the places we'd want to get collect stats, and when subscribe to the notifications wherever we want. This way we don't have to hardcode the stats into the gem. This also adds a Noop instrumenter class for when an instrumenter is not passed in.