add the ability to optionally send urls
We don't want to send sensitive data to lightstep, but sometimes the urls of faraday requests contain sensitive information. We want to able to disable the sending of those urls. This also adds an `http.domain` tag, which will at least give us the domain of the request even if the url needs to be scrubbed.