add the error:true tag to spans with error status codes
We add the error:true tag to requests that have error status codes in other services like GLB, but haven't done the same for the app requests recorded via hubstep. This updates both middlewares to include that error:true tag when the status code is equal to or higher than 400.