Support distributed tracing
This adds a new child_of keyword argument to Tracer#span so that extracted
tracing information can be injected. Using the extract method, one can get
extract a SpanContext from a tracing carrier (typically, HTTP headers) and
connect a remote service to its caller as a child. This enables better ordering
of operations in the LightStep UI.